Cleaned up some the rate limited yo-variables.
Also changed RateLimitedYoFramePose to use a YoFrameQuaternion instead
of a YoFrameOrientation.

Changed RateLimitedYoFramePose to create less YoVariables.

I'm not really happy about having to duplicate RateLimitedYoFrameVector
to create RateLimitedYoFramePoint. That should be easy to cleanup later
though.

Created AlphaBasedOnBreakFrequencyProvider.

Changed the alpha filter to be compatible with parameter.

Propagated the change to the classes using the AlphaFilteredYoVariable.

Made the same change to the alpha filtered yo frame quaternion.

Forgot to create alpha when not provided.

Updated codebase.